Moved TranslationFactor prototype into sfm directory
